작업용

+++ src/main/java/com/test/bidon/service/ReviewBoardService.java
@@ -17,7 +17,7 @@ public class ReviewBoardService {
     private ReviewBoardRepository reviewBoardRepository;
 
     // ReviewBoardId로 ReviewBoard 데이터 가져오기
-    public ReviewBoard getReviewBoardById(Long reviewBoardId) {
+    public ReviewBoard getReviewBoardById(Integer reviewBoardId) {
         return reviewBoardRepository.findById(reviewBoardId).orElse(null);
     }
